Created requests package and ExecuteRequest class
